+++ changelog.md
@@ -5,7 +5,23 @@ All notable changes will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.1] - 2022-10-15
+## [0.1.2] - 2022-12-13
+### Added
+
+- Compatibility with Linux for the DataManager
+- New UIactions: Play_music, stop_music
+- Notify Screen
+- CheckButtons
+
+### Fixed
+
+- Minor Bugs
+
+### Changed
+
+- ImageButtons Antialiasing control
+
+## [0.1.1] - 2022-11-15
 ### Added
 
 - ImageButtons